Removed unnecessary comments, added custom functions and if's to js file
Some comments were unnecessary (commenting on lists), so they were
removed. In the javascript file, created custom function that resets
the display when you click on a nav item, so that it could be called in
each nav function, instead of the ~10 lines being rewritten each time.
Then added conditionals for each nav click, so that if a user is
already on ie. music, and they click the music nav again, it won't
reload the entire section and do the animation all over again. Also
added some more comments to js.